==> Step 04 - Get plain text from TLG and filter.ipynb <==
This notebook is responsible for extracting the plain-text of the TLG corpus and filtering texts according to various little filters, **mostly to exclude poetry, fragments and uncertain texts**.

# Documenting
from typing import Generator

# OS
import glob
import os.path

# Data
import csv
import lxml.etree as ET
import pandas
from collections import Counter

# Operations
import regex as re
import unicodedata

# UI
import tqdm

## Constants

#PATH = "/home/thibault/Downloads/comp/Diogenes-Resources/xml/tlg/tlg*.xml"
PATH = "patres/tlg*.xml"
MIN_SIZE = 1000 # Number of first words to take
ForbiddenTGs = {
    "tlg0527", "tlg0031"
}

## XML parsing

def normalize(text: str) -> str:
    return unicodedata.normalize("NFKC", re.sub(r"\s{2,}|\n+", " ", text))

cnt = re.compile("\w+")

NS = {"namespaces": {"t": "http://www.tei-c.org/ns/1.0"}}

XSL = """<xsl:stylesheet xmlns:xsl="http://www.w3.org/1999/XSL/Transform"
    xmlns:xs="http://www.w3.org/2001/XMLSchema"
    xmlns:t="http://www.tei-c.org/ns/1.0"
    exclude-result-prefixes="xs"
    version="1.0">
    <xsl:output method="text"/>
    <xsl:template match="t:TEI">
        <xsl:apply-templates select="//t:body" />
    </xsl:template>
    <xsl:template match="t:body">
        <xsl:apply-templates/>
    </xsl:template>
    <xsl:template match="t:label"/>
    <xsl:template match="t:note"/>
    <xsl:template match="t:head"/>
</xsl:stylesheet>"""

XSL = ET.XSLT(ET.fromstring(XSL))

## Filtering functions

def forbidden_author(string) -> bool:
    string = string.lower()
    if "epistula" in string:
        return True
    if "comment" in string:
        return True
    if "scholia" in string:
        return True
    if "lexic" in string:
        return True
    if "acta" in string:
        return True
    if "fragm" in string:
        return True
    if "corpus" in string:
        return True
    if "anonym" in string:
        return True
    if "pseud" in string:
        return True
    if "testam" in string:
        return True
    if "vita" in string:
        return True
    if string.endswith(" et"):
        return True
    if "evangeli" in string:
        return True
    if "floril" in string:
        return True
    if "anthologia" in string:
        return True
    if "paraphras" in string:
        return True
    if "apocal" in string:
        return True
    if "certamen" in string:
        return True
    if "comica adespota" in string:
        return True
    if "commentaria" in string:
        return True
    if "concilia" in string:
        return True
    if "epica adespota" in string:
        return True
    if "etymologicum" in string:
        return True
    if "epistula ecclesiarum" in string:
        return True
    if "gnologium" in string:
        return True
    if "lyrica adespota" in string:
        return True
    if "proverbi" in string:
        return True
    if "oracula" in string:
        return True
    if "periplus" in string:
        return True
    if "incertus" in string:
        return True
    if "socraticorum epistulae" in string:
        return True
    if "socrat" in string:
        return True
    if "[" in string:
        return True
    if "anonym" in string:
        return True
    if "pseudo" in string:
        return True
    if "ps." in string:
        return True
    if "historia" in string:
        return True
    
def is_poetry(xml):
    return len(xml.xpath("//t:l", **NS)) > 0

def has_fragment(xml):
    return len(xml.xpath("//t:div[@type='fragment']", **NS)) > 0

## Data wrangling functions

def get_tg(filename):
    return os.path.basename(filename)[:7]

def rename_author(author):
    return re.sub(r"(\s+\w+\.)", "", author).replace(" et", "")

## Data accumulation

# Output Data
data = []

# Count the number of files we have per textgroup
tgs = Counter([
    get_tg(file)
    for file in glob.glob(PATH)
])

# Statistics
passed = 0
tgcount = Counter()
ignored_authors = set()
    
for file in tqdm.tqdm(glob.glob(PATH)):
    try:
        xml = ET.parse(file)
        tg = get_tg(file)
        author = str(xml.xpath("/t:TEI/t:teiHeader/t:fileDesc/t:titleStmt/t:author/text()", **NS)[0])
        title = xml.xpath("/t:TEI/t:teiHeader/t:fileDesc/t:titleStmt/t:title/text()", **NS)[0]
        
        if tg in ForbiddenTGs:
            passed += 1
            continue
        elif forbidden_author(author):
            ignored_authors.add(author)
            continue
        elif is_poetry(xml):
            #print(f"Ignoring {file} for poetry reason")
            ignored_authors.add(author)
            continue
        elif has_fragment(xml):
            #print(f"Ignoring {file} for fragment reason")
            ignored_authors.add(author)
            continue
            
        rawtext = normalize(str(XSL(xml))).strip()
        tokens = cnt.findall(rawtext)
        
        if len(tokens) < MIN_SIZE:
            passed += 1
            continue

        data.append({
            "file": os.path.basename(file)[:-4],
            "orig_author": author,
            "author": rename_author(str(author)),
            "title": str(title),
            "textgroup": tg,
            "tokens": len(tokens),
            "full-text-raw": rawtext
        })
        tgcount[tg] += 1
    except Exception as E:
        raise E
        passed += 1
        print(f"Failing on {file}")
        continue
print(passed)
print(ignored_authors)

## Exporting

df = pandas.DataFrame(data)
print("Before filtering on Title", df.shape)
df = df[~df.title.str.contains("Dub\.|Sp\.|Fragm|Excerpt|(e cod\.)|Suda|recensio|fragm|sp\.|dub\.|(fort\. auctore)|Scholia")]
print("After filtering on Title", df.shape)


df.to_csv("tlg-texts.csv", index=False)
df.head()

for x in sorted(df.orig_author.unique()):
    print(x)

## Lemmatization

Requires to have set-up `bert-env` virtual environment.

!./bert-env/bin/python tag-in-xml.py tlg-texts.csv
