==> Step 05 - Build MFW.ipynb <==
This notebooks build the most frequent words lists.

import os
import glob
import json
import pandas as pd
from collections import Counter
from typing import List
import tqdm

def get_affixes(token):
    if len(token) <= 3:
        yield token
    if len(token) >= 2:
        yield f"_{token[:2]}"
        yield f"{token[-2:]}_"
    if len(token) > 3:
        yield f"{token[:3]}"
        yield f"{token[-3:]}"
        

texts = pd.read_csv("tlg-texts.csv").file

def read_tsv(file):
    for line in file:
        yield line.split()[:2]

        
def get_trigrams(tokens: List[str]) -> List[str]:
    return ["-".join(tokens[i:i+3]) for i in range(len(tokens)-3+1)]


tokens = Counter()
poses = Counter()
trigrams = Counter()

for text in tqdm.tqdm(texts):
    path = f"./tagged/{text}-tagged.txt"
    if not os.path.exists(path):
        continue
    with open(path) as f:
        doc = list(read_tsv(f))
        tokens += Counter([
            token.lower()
            for token, pos in doc
            if pos[0] != "u" and pos[0] != "n"
        ])
        poses += Counter(get_trigrams([pos[0] for _, pos in doc if pos[0] != "u"]))
        trigrams += Counter([
            t
            for token, pos in doc
            for t in get_affixes(token.lower())
            if pos[0] != "u" 
        ])
        
        
with open("mfw.json", "w") as f:
    json.dump(tokens.most_common(2000), f)

with open("mfp.json", "w") as f:
    json.dump(poses.most_common(2000), f)
    
with open("mft.json", "w") as f:
    json.dump(trigrams.most_common(4000), f)

trigrams.most_common(2000)
