==> Step 08 - Siamese Sequential.ipynb <==
import random
import re

import pandas as pd
import pytorch_lightning as pl

from sklearn.metrics import roc_auc_score

import seaborn as sns
import matplotlib.pyplot as plt

from freestyl.dataset.dataframe_wrapper import DataframeWrapper
from freestyl.supervised.siamese import train_dataframewrappers
from freestyl.supervised.siamese import get_df_prediction
from freestyl.supervised.siamese.utils import score_from_preds
from freestyl.utils import plot_aucroc_curve
#>>> x = [1, 2, 3, 4, 5, 6]
#>>> .shuffle(x)

NEW_DATASET = False

seed = 42
IGNORE_KEYS = [
    "file", "author", "textgroup", "title", "tokens", "length", "modified_text"
]
REMOVED = ["Euclides"]


## Import Data

def get_train_dev_test(filtre, seed=42, ratio=10):
    test = len(filtre) * ratio // 100
    dev_and_test = test * 2
    print(f"Train categorical samples: {len(filtre)-dev_and_test}")
    print(f"Dev   categorical samples: {dev_and_test-test}")
    print(f"Test  categorical samples: {test}")
    #r = random.Random(seed)
    #r.shuffle(filtre)
    return filtre[:-dev_and_test], filtre[-dev_and_test:-test], filtre[-test:]


if NEW_DATASET:
    df = pd.read_csv("tlg-features.csv").sample(frac=1)
    #df = df[df.tokens >= 5000]
    #print(">= 5000", df.shape)
    #df = df[~df.file.isin(POETRY)]
    #print(">= Poetry", df.shape)
    # df = df[~df.title.str.contains("Dub\.|Sp\.|Fragm|Excerpt|(e cod\.)|Suda|recensio|fragm|sp\.|dub\.|(fort\. auctore)|Scholia")]
    print("Title filter", df.shape)
    #df = df[~df["full-pos-text"].isna()]
    #print(">= POS missing", df.shape)
    
    # Filter based on authors, to generalize better. Everything should be out of domain\
    train, dev, test = [], [], []
    if SPLIT_ON_AUTHORS:
        authors = df.author.value_counts()
        a, b, c = get_train_dev_test(authors[authors > 1].index.tolist())
    else:
        a, b, c = get_train_dev_test(df.index.tolist())
        
    train.extend(a)
    dev.extend(b)
    test.extend(c)
    
#     a, b, c = get_train_dev_test(authors[authors == 1].index.tolist())
#     train.extend(a)
#     dev.extend(b)
#     test.extend(c)

    if SPLIT_ON_AUTHORS:
        train = df[df.author.isin(train)].copy(deep=True)
        dev = df[df.author.isin(dev)].copy(deep=True)#.author.value_counts()
        test = df[df.author.isin(test)].copy(deep=True)#.author.value_counts()
    else:
        train = df[df.index.isin(train)].copy(deep=True)
        dev = df[df.index.isin(dev)].copy(deep=True)#.author.value_counts()
        test = df[df.index.isin(test)].copy(deep=True)#.author.value_counts()
        
    train.to_csv("tlg-train.csv", index=False)
    dev.to_csv("tlg-dev.csv", index=False)
    test.to_csv("tlg-test.csv", index=False)
else:
    train = pd.read_csv("tlg-train.csv")
    dev = pd.read_csv("tlg-dev.csv")
    test = pd.read_csv("tlg-test.csv")
    
train = train[~train.author.isin(REMOVED)]
dev = dev[~dev.author.isin(REMOVED)]
test = test[~test.author.isin(REMOVED)]

print(f"Train Shape : {train.shape}")
print(f"Dev Shape : {dev.shape}")
print(f"Test Shape : {test.shape}")

## Util functions

def assign_normalization(dfw):
    dfw.normalized._dataframe = dfw.dataframe.fillna(0)
    
def get_scores(scores, distance: float) -> None:
    scores["Attribution"] = scores.Distance < distance
    tp = scores[scores.IsAPair].Attribution.sum() 
    fn = scores[scores.IsAPair].Attribution.shape[0] - tp
    fp = scores[~scores.IsAPair].Attribution.sum()
    print(f"True positives: {tp}\nFalse Negative {fn}")
    print(f"False positives: {fp}\nAccuracy: {tp/(fn+tp):.2f}")

## Automatically retrieve some constant parameters

POS_COLS = [
    col
    for col in train.columns
    if col.startswith("$POS$")
]
FW_COLS = [
    col
    for col in train.columns
    if col.startswith("$MFW$")
]
IGNORE = set(IGNORE_KEYS + POS_COLS + FW_COLS)

## Get DataFrameWrapper

data = DataframeWrapper(train, target="author", label=["author", "title"], x_ignore=IGNORE)
assign_normalization(data)
data_dev = DataframeWrapper(dev, target="author", label=["author", "title"], x_ignore=IGNORE)
assign_normalization(data_dev)
data_test = DataframeWrapper(test, target="author", label=["author", "title"], x_ignore=IGNORE)
assign_normalization(data_test)

## CHecking some details

data.xs.head()

## Training

models, trainer = train_dataframewrappers(
    train=data,
    dev=data_dev,
    test=data_test,
    learning_rate=1e-3,
    margin=1,
    mode="sequential",
    loss="linearManhattan",
    sample=None,
    batch_size=128,
    gpus=1,
    accelerator="gpu",
    min_epochs=100,
    patience=10,
    sequential_min_token_freq=5,
    document_hidden_size = 128,
    embedding_size = 100,
    dropout=.30,
    miner_for_dev=True,
    sequential_model="AttentionalGRU",
    sequential_text_key="modified_text"
)


models

## Evaluating Dev for Test Threshold

trainer = pl.Trainer(
    gpus=1,
    accelerator="gpu"
)

scores = get_df_prediction(trainer, model=models, compared=data_dev)
#scores
print(f"ROC: {roc_auc_score(scores.IsAPair, scores.Probability)}")

scores[scores.Distance < 1].Distance.plot.box()
scores.head()

## Study classes

for cls in scores.ComparedClass.unique():
    plt.figure()
    sns.boxplot(data=scores[(scores.ComparedClass==cls) & (scores.Distance<1)], x="IsAPair", y="Distance")
    plt.gca().set_title(cls)
    #scores.groupby("ComparedClass").plot.box(y="Distance", x="IsAPair")

DEV_PROBA = 0

MAX_DISTANCE = .75
get_scores(scores, MAX_DISTANCE)
print("\n===\nWithout sampling\n===\n")
get_scores(scores[(scores.ComparedLabel != scores.ComparatorLabel)], MAX_DISTANCE)

plot_aucroc_curve(scores.IsAPair, scores.Probability, nth=1000)
ax = plt.gca()
ax.legend(bbox_to_anchor=(1.1, 1.05))

## Evaluating Test with Dev Threshold

models.eval()
MAX_DISTANCE = .75

pairs = get_df_prediction(trainer, model=models, compared=data_test, threshold=MAX_DISTANCE)

print(f"ROC: {roc_auc_score(pairs.IsAPair, pairs.Probability)}")
MAX_DISTANCE = .75
get_scores(scores, MAX_DISTANCE)
print("\n===\nWithout sampling\n===\n")
get_scores(scores[(scores.ComparedLabel != scores.ComparatorLabel)], MAX_DISTANCE)

plot_aucroc_curve(pairs.IsAPair, pairs.Probability, nth=1000)
ax = plt.gca()
ax.legend(bbox_to_anchor=(1.1, 1.05))

pairs.to_csv("test-results.csv")

## On Voicu !

import regex as re
import json
import unicodedata

df = pd.read_csv("pc-features.csv")
DFW = DataframeWrapper(df, label=("author", "title"), target="title", 
                       x_ignore=IGNORE.union({'$MFW$μηδ', '$MFW$εὖ', '$MFW$ἐς', '$MFW$ταὐτὸν', '$MFW$ἧττον', '$MFW$μάλα', '$MFW$οὔτ', '$MFW$τ', '$MFW$γ', '$MFW$ἤτοι', '$MFW$ἥ', '$MFW$εἴπερ', '$MFW$η', '$MFW$κ', '$MFW$α', '$MFW$β', '$MFW$αʹ', '$MFW$[', '$MFW$p', '$MFW$̣', '$MFW$𐅻'}))
#
print(len(DFW.features))
DFW.update_features(data.features)
assign_normalization(DFW)
#DFW._features = data.features
print(len(data.features))
print(len(DFW.features))

models.eval()
pairs = get_df_prediction(trainer, model=models, compared=DFW, threshold=1)
pairs = pairs[pairs.ComparedLabel != pairs.ComparatorLabel]

#pairs.to_csv("pairs-last-experiment.csv")

pairs.Attribution = pairs.Distance < .85
print(f"Pairing: {pairs.Attribution.sum()/len(pairs)}")
pairs[pairs.Distance < .85]

## Add FP and TP for each distance based on test
